Non-uniqued NSValueTransformers
This exposes an interface for subclasses to provide NSValueTransformers for properties.
There are _no_ default transformers and no (public) interface to automatically convert NSDates and the like.

+++ example/FCModelTest/AppDelegate.m
@@ -116,7 +116,7 @@
     NSMutableSet *colorsUsedAlready = [NSMutableSet set];
     
     // Put some data in the table if there's not enough
-    int numPeople = [Person numberOfInstances];
+    NSUInteger numPeople = [Person numberOfInstances];
     while (numPeople < 26) {
         Person *p = [Person new];
         do {
